se cambia tipo de dato de dni
se cambia de long a long long y se corrige el bug de los 11 digitos.
Ademas se agregan las validaciones que realizo Yami para contar los digitos.

+++ PersonaManager.cpp
@@ -6,7 +6,7 @@
 
 
 void PersonaManager::CargarPersona() {
-    long dni;
+    long long dni;
     std::string nombre, apellido;
     Fecha fechaN;
